+++ src/Console/Database/SeedCommand.php
@@ -17,7 +17,7 @@ class SeedCommand extends BaseSeedCommand
     protected function getOptions()
     {
         return array_merge(parent::getOptions(), [
-            ['layer', null, InputOption::VALUE_OPTIONAL, 'Слой к которому применяется команда.', 'main'],
+            ['layer', null, InputOption::VALUE_OPTIONAL, 'Слой к которому применяется команда.', null],
         ]);
     }