+++ src/Console/Migrations/MigrateMakeCommand.php
@@ -24,5 +24,5 @@ class MigrateMakeCommand extends BaseMigrateMakeCommand
         {--path= : The location where the migration file should be created}
         {--realpath : Indicate any provided migration file paths are pre-resolved absolute paths}
         {--fullpath : Output the full path of the migration}
-        {--layer=main : Слой к которому применяется команда.}';
+        {--layer= : Слой к которому применяется команда.}';
 }